Moved tests to PreviouslyDownloadedOnlyTest
